+++ crates/rust-client/src/store/sqlite_store/notes.rs
@@ -70,7 +70,8 @@ struct SerializedOutputNoteParts {
 // ================================================================================================
 type NoteQueryParams = Vec<Rc<Vec<Value>>>;
 impl NoteFilter {
-    /// Returns a [String] containing the query for this Filter
+    /// Returns a [String] containing the full output notes query for this Filter and a vector of
+    /// parameters to be used in it.
     fn to_query_output_notes(&self) -> (String, NoteQueryParams) {
         let base = "SELECT
                     note.recipient_digest,
@@ -86,6 +87,8 @@ impl NoteFilter {
         (query, params)
     }
 
+    /// Returns a [String] containing the output notes query conditions for this Filter and a vector
+    /// of parameters to be used in it.
     fn output_notes_condition(&self) -> (String, NoteQueryParams) {
         let mut params = Vec::new();
         let condition = match self {
@@ -107,7 +110,7 @@ impl NoteFilter {
                     OutputNoteState::STATE_EXPECTED_FULL
                 )
             },
-            NoteFilter::Processing => "1 = 0".to_string(), // There are no processing output notes
+            NoteFilter::Processing | NoteFilter::Unverified => "1 = 0".to_string(), /* There are no processing or unverified output notes */
             NoteFilter::Unique(note_id) => {
                 let note_ids_list = vec![Value::Text(note_id.inner().to_string())];
                 params.push(Rc::new(note_ids_list));
@@ -131,7 +134,6 @@ impl NoteFilter {
                 params.push(Rc::new(nullifiers_list));
                 "note.nullifier IN rarray(?)".to_string()
             },
-            NoteFilter::Unverified => "1 = 0".to_string(), // there are no unverified output notes
             NoteFilter::Unspent => {
                 format!(
                     "state_discriminant in ({}, {})",
@@ -144,6 +146,8 @@ impl NoteFilter {
         (condition, params)
     }
 
+    /// Returns a [String] containing the full input notes query conditions for this Filter and a
+    /// vector of parameters to be used in it.
     fn to_query_input_notes(&self) -> (String, NoteQueryParams) {
         let base = "SELECT
                 note.assets,
@@ -162,6 +166,8 @@ impl NoteFilter {
         (query, params)
     }
 
+    /// Returns a [String] containing the input notes query conditions for this Filter and a vector
+    /// of parameters to be used in it.
     fn input_notes_condition(&self) -> (String, NoteQueryParams) {
         let mut params = Vec::new();
         let condition = match self {
@@ -268,7 +274,6 @@ impl SqliteStore {
         filter: NoteFilter,
     ) -> Result<Vec<OutputNoteRecord>, StoreError> {
         let (query, params) = filter.to_query_output_notes();
-
         let notes = self
             .db()
             .prepare(&query)?