+++ crates/web-client/src/sync.rs
@@ -1,14 +1,14 @@
 use wasm_bindgen::prelude::*;
 
-use crate::WebClient;
+use crate::{models::sync_summary::SyncSummary, WebClient};
 
 #[wasm_bindgen]
 impl WebClient {
-    pub async fn sync_state(&mut self) -> Result<JsValue, JsValue> {
+    pub async fn sync_state(&mut self) -> Result<SyncSummary, JsValue> {
         if let Some(client) = self.get_mut_inner() {
             let sync_summary = client.sync_state().await.unwrap();
 
-            Ok(JsValue::from_f64(sync_summary.block_num as f64))
+            Ok(sync_summary.into())
         } else {
             Err(JsValue::from_str("Client not initialized"))
         }